Create dir for Uploads
Will create a "Uploads" directory to add songs if it doesn't exist. Will successfully upload the file.

**  NOTE in order for the upload to work you may have to make changes to the PHP.ini file in Xampp!

Find and increase values for the following:
.upload_max_filesize
.post_max_size

+++ AddRecord.php
@@ -102,9 +102,25 @@
 						$songLength, $songTitle,
 						$songYear, $songFile);
 				
-				// Upload Song file to "Upload" folder
-				move_uploaded_file($temp, "upload/".$songFile);
-				echo "<p>\"$songTitle\" has been successfully uploaded</p>";
+				$dirName = "Upload";	// Directory to create
+				
+				// If the directory doesn't exist, create it
+				if(!is_dir($dirName))	
+					mkdir("$dirName");	
+				
+				if(is_dir($dirName))	// If Upload directory exists
+				{
+					try	// Try to upload to Upload folder
+					{
+						// Upload Song file to "Upload" folder
+						move_uploaded_file($temp, "$dirName/".$songFile);
+						echo "<p>\"$songTitle\" has been successfully uploaded</p>";
+					} 
+					catch(exception $e){
+						echo "<p>$e</p>";
+					}
+					
+				}
 			}
 		}
 		else
